pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "tracksuite"
description = "ecflow suite tracking and deploying toolkit"
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
  { name = "Corentin Carton de Wiart", email = "corentin.carton@ecmwf.int" },
]
license = { text = "Apache License Version 2.0" }
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: Unix",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]
dynamic = ["version", "readme"]

dependencies = [
    "gitpython >= 3.1.25"
]

[project.optional-dependencies]
test = ["pytest", "mocker", "pytest-mock"]

[project.urls]
"Source code" = "https://github.com/ecmwf/tracksuite"

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests"]

[project.scripts]
    tracksuite-init = "tracksuite.init:main"
    tracksuite-deploy = "tracksuite.deploy:main"
    tracksuite-revert = "tracksuite.revert:main"
    tracksuite-update-defs = "tracksuite.definition:main"
    tracksuite-replace = "tracksuite.replace:main"

[tool.setuptools_scm]
write_to = "tracksuite/_version.py"
write_to_template = '''
# Do not change! Do not track in version control!
__version__ = "{version}"
'''
parentdir_prefix_version='tracksuite-'  # get version from GitHub-like tarballs
fallback_version='0.3.3'
